flask-react-app/backend/requirements.txt
# Backend
flask
flask_cors
flask_socketio
python-dotenv
pandas
numpy
openpyxl
requests
open3d
matplotlib
scipy
opencv-python
mysql-connector-python
